requirements.txt
# Core packages
boto3==1.36.5
botocore>=1.27.0
python-dateutil>=2.8.0

# HTML processing
beautifulsoup4==4.12.3
lxml==5.3.0

# HTTP requests for Cloudflare API
requests>=2.25.0
urllib3>=1.26

# Date/timezone handling
pytz==2024.2

# Text processing
ftfy==6.3.1

# Additional dependencies for profile scraping
python-dotenv>=0.19.0
